panacloud_assistant/panacloud_agent/pyproject.toml
[project]
name = "panacloud-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "graphviz>=0.20.3",
    "openai-agents[litellm,viz]>=0.0.14",
    "python-dotenv>=1.1.0",
]
